requirements.txt
astroid==2.3.3
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
COVID19Py==0.3.0
cryptography==2.9
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
pylint==2.4.4
pyOpenSSL==19.1.0
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
python-dotenv==0.12.0
requests==2.23.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2